Implement Naive Bayes classifier
This classifier is built on the conditional probabilities computed
previously. Since it assumes all attributes are conditionally
independet, we can compute the class-conditional probability by simply
multiplying the probabilities computed previously.

This also adds an accuracy method that computes the accuracy of the
provided labeling. This should be able to be reused for later parts
doing cross-validation.